+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Development
 -----------
 
+* Placeholder
+
+2.8.6
+-----
+
 * Fix issue with `get_reporting_data` and `get_baseline_data` when passing data with non-UTC timezones.
 
 2.8.5